pyproject.toml
[project]
name = "mindmap-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=24.1.0",
    "aiolimiter>=1.2.1",
    "aiosqlite>=0.21.0",
    "anthropic>=0.49.0",
    "async-timeout>=5.0.1",
    "fuzzywuzzy>=0.18.0",
    "google-genai>=1.5.0",
    "numpy>=2.2.3",
    "openai>=1.65.5",
    "psutil>=7.0.0",
    "python-decouple>=3.8",
    "python-levenshtein>=0.27.1",
    "ruff>=0.9.10",
    "spacy>=3.8.4",
    "sqlalchemy>=2.0.38",
    "sqlmodel>=0.0.24",
    "termcolor>=2.5.0",
    "tiktoken>=0.9.0",
    "transformers>=4.49.0",
]
